Add programs for the implementation of laboratory work number 6
In the CodeFirstApp folder there is an implementation of the database by the code first approach, respectively there is implemented crud news + ui for this program and a migration database is created.
In the folder DatabaseFirstApp there is an implementation of the database using the database first approach, respectively there is implemented crud news + ui for this program.